requirements.txt
# requirements.txt
# -----------------
streamlit  
openai==1.93.0             # OpenAI Python SDK :contentReference[oaicite:1]{index=1}
PyPDF2==3.0.1
fpdf2==2.8.3
python-docx==1.1.2            # fast PDF text extraction (import as “fitz”) :contentReference[oaicite:2]{index=2}
requests==2.32.4           # HTTP client :contentReference[oaicite:3]{index=3}
typing-extensions==4.14.1  # back-ports for newer typing features :contentReference[oaicite:4]{index=4}
beautifulsoup4==4.12.3
types-beautifulsoup4
python-dotenv
python-dateutil==2.9.0.post0
types-python-dateutil
types-fpdf2
sentence-transformers
faiss-cpu
plotly
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl


# --- development / CI ---
pre-commit==4.2.0          # hook runner :contentReference[oaicite:5]{index=5}
pytest              # test runner :contentReference[oaicite:6]{index=6}
black==25.1.0              # code formatter :contentReference[oaicite:7]{index=7}
isort==6.0.1               # import sorter :contentReference[oaicite:8]{index=8}
flake8==7.3.0              # linter/quality checks :contentReference[oaicite:9]{index=9}
types-requests==2.32.4.20250611  # typing stubs for “requests” :contentReference[oaicite:10]{index=10}
lxml==5.2.1
pdfplumber
